made assembly work w minigame class
Changed "player" signal name to "player_signal" so it doesn't conflict w minigame class player.
You have to connect the interaction w the minigame and set the stuff in inspector for it to work, and it only runs on main.
Change it back to extends node2d to run it on its own.